+++ examples/platform/silabs/SiWx917/SiWx917/rsi_if.c
@@ -422,6 +422,7 @@ static void wfx_rsi_save_ap_info()
 static void wfx_rsi_do_join(void)
 {
     int32_t status;
+    rsi_security_mode_t connect_security_mode;
 
     if (wfx_rsi.dev_state & (WFX_RSI_ST_STA_CONNECTING | WFX_RSI_ST_STA_CONNECTED))
     {
