+++ src/platform/silabs/SiWx917/BUILD.gn
@@ -43,8 +43,11 @@ static_library("SiWx917") {
     "${silabs_platform_dir}/DiagnosticDataProviderImpl.cpp",
     "${silabs_platform_dir}/DiagnosticDataProviderImpl.h",
     "${silabs_platform_dir}/InetPlatformConfig.h",
+    "${silabs_platform_dir}/KeyValueStoreManagerImpl.cpp",
     "${silabs_platform_dir}/KeyValueStoreManagerImpl.h",
     "${silabs_platform_dir}/Logging.cpp",
+    "${silabs_platform_dir}/MigrationManager.cpp",
+    "${silabs_platform_dir}/MigrationManager.h",
     "${silabs_platform_dir}/PlatformManagerImpl.h",
     "${silabs_platform_dir}/SilabsConfig.cpp",
     "${silabs_platform_dir}/SilabsConfig.h",
@@ -53,7 +56,6 @@ static_library("SiWx917") {
     "../../SingletonConfigurationManager.cpp",
     "BLEManagerImpl.cpp",
     "ConfigurationManagerImpl.cpp",
-    "KeyValueStoreManagerImpl.cpp",
     "PlatformManagerImpl.cpp",
     "bluetooth/rsi_ble_config.h",
     "bluetooth/wfx_sl_ble_init.c",
